requirements.txt
blinker==1.4
boto3==1.17.109
botocore==1.20.109
certifi==2021.5.30
click==8.0.1
Flask==2.0.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
python-dateutil==2.8.1
python-dotenv==0.18.0
s3transfer==0.4.2
sentry-sdk==1.3.0
six==1.16.0
urllib3==1.26.6
Werkzeug==2.0.1
